Criada a pasta util, os arquivos de validacao das entidadaes e o arquivo que recebe os erros da requisicao
