+++ src/data/usuarios.json
@@ -16,5 +16,17 @@
         "nome": "Marcio",
         "email": "marcio.com",
         "senha": "teste1"
+    },
+    {
+        "id": 4,
+        "nome": "Fulano",
+        "email": "fulano@gmail.com",
+        "senha": "fulano"
+    },
+    {
+        "id": 5,
+        "nome": "Beltrano",
+        "email": "beltrano@gmail.com",
+        "senha": "beltrano"
     }
 ]
